refactor(module): provide tailwind config through template
