+++ src/runtime/utils/colors.ts
@@ -1,4 +1,4 @@
-import { omit } from './runtime/utils/lodash'
+import { omit } from './lodash'
 import { kebabCase, camelCase, upperFirst } from 'scule'
 
 const colorsToExclude = [
